+++ src/index.js
@@ -1,5 +1,6 @@
 import express from "express";
 import { ValidationError } from "express-validation";
+import { validationResult } from "express-validator";
 import routes from "./routes/index.js";
 import { errResponse } from "./utils/response.js";
 import statusCode from "./utils/statusCode.js";
@@ -12,7 +13,9 @@ app.use(express.urlencoded({ extended: false }));
 app.use("/", routes);
 // eslint-disable-next-line
 app.use((err, req, res, next) => {
-  if (err instanceof ValidationError) {
+  const errors = validationResult(req);
+
+  if (err instanceof ValidationError || !errors.isEmpty()) {
     return res
       .status(statusCode.BAD_REQUEST)
       .send(errResponse(statusCode.BAD_REQUEST, message.BAD_REQUEST));
