+++ src/routes/index.js
@@ -1,17 +1,14 @@
 import { Router } from "express";
-import sewerLevelRouter from "./sewerLevelRoute.js";
+import { query } from "express-validator";
+import controller from "../controllers/sewerLevelController.js";
 
 const router = Router();
 
-const defaultRoutes = [
-  {
-    path: "/",
-    route: sewerLevelRouter,
-  },
-];
-
-defaultRoutes.forEach((route) => {
-  router.use(route.path, route.route);
-});
+router.get(
+  "/",
+  query("startDate").isISO8601().toDate(),
+  query("endDate").isISO8601().toDate(),
+  controller
+);
 
 export default router;
